Moved database to infra folder
