--- services/assistant/app/database/sql_db/__init__.py
@@ -1,3 +1,0 @@
-from app.database.sql_db.orm import get_db_factory
-
-__all__ = ["get_db_factory"]
