--- services/assistant/app/database/sql_db/repositories/__init__.py
@@ -1,5 +1,0 @@
-from app.database.sql_db.repositories.booking_repository import SQLBookingRepository
-from app.database.sql_db.repositories.user_repository import SQLUserRepository
-from app.database.sql_db.repositories.xray_repository import SQLXrayRepository
-
-__all__ = ["SQLBookingRepository", "SQLUserRepository", "SQLXrayRepository"]
