+++ services/assistant/app/infra/database/sql_db/__init__.py
@@ -1,0 +1,3 @@
+from app.infra.database.sql_db.orm import get_db_factory
+
+__all__ = ["get_db_factory"]
