+++ services/assistant/app/infra/database/sql_db/repositories/__init__.py
@@ -1,0 +1,7 @@
+from app.infra.database.sql_db.repositories.booking_repository import (
+    SQLBookingRepository,
+)
+from app.infra.database.sql_db.repositories.user_repository import SQLUserRepository
+from app.infra.database.sql_db.repositories.xray_repository import SQLXrayRepository
+
+__all__ = ["SQLBookingRepository", "SQLUserRepository", "SQLXrayRepository"]
