+++ services/assistant/app/infra/database/sql_db/repositories/base_repository.py
