+++ services/assistant/manage.py
@@ -6,9 +6,11 @@ from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker
 from settings import settings, logger
 from app.app import AppFactory
 from app.adapters.bot import BotFactory, BotServicesFactory, BotSettings
-from app.database.sql_db import get_db_factory
-from app.database.sql_db.repositories.repository_factory import SQLRepositoryFactory
-from app.database.sql_db.repositories import (
+from app.infra.database.sql_db import get_db_factory
+from app.infra.database.sql_db.repositories.repository_factory import (
+    SQLRepositoryFactory,
+)
+from app.infra.database.sql_db.repositories import (
     SQLBookingRepository,
     SQLUserRepository,
     SQLXrayRepository,
